Changes to be committed: 	new file:   assets/icons/ui/battle-gear.svg 	new file:   assets/icons/ui/battle-gear.svg.import 	new file:   assets/icons/ui/knapsack.svg 	new file:   assets/icons/ui/knapsack.svg.import 	modified:   source/ui/buttons/Buttons.tscn 	modified:   source/ui/ui_grid/Grid.gd

+++ source/ui/ui_grid/Grid.gd
@@ -6,14 +6,14 @@ func _ready() -> void:
 	hide()
 	
 func conf(_b : bool):
-	if _b:
+	if _b and !visible:
 		$Tween.remove_all()
-		$Tween.interpolate_property(self, "self_modulate:a", 0, 1, 0.2, Tween.TRANS_BACK, Tween.EASE_IN)
+		$Tween.interpolate_property(self, "self_modulate:a", self_modulate.a, 1, 0.2, Tween.TRANS_BACK, Tween.EASE_IN)
 		show()
 		$Tween.start()
-	else:
+	if !_b and visible:
 		$Tween.remove_all()
-		$Tween.interpolate_property(self, "self_modulate:a", 1, 0, 0.2, Tween.TRANS_BACK, Tween.EASE_IN)
+		$Tween.interpolate_property(self, "self_modulate:a", self_modulate.a, 0, 0.2, Tween.TRANS_BACK, Tween.EASE_IN)
 		$Tween.start()
 		yield($Tween,"tween_all_completed")
 		hide()
